Teach the UserList to track idents and hostnames
Previously, the UserList would just be a list of strings representing
the nicknames of people in the channel, made possibly by parsing the
RPL_NAMREPLY (353).

This has been changed to include idents and hostnames too (in
nick!ident@hostname). This sort of information is not included in the
RPL_NAMREPLY (353) messages. To obtain this information, RPL_WHOREPLY
(353) is required. 'NAMES' has been changed to 'WHO' and NamesObserver
is now WhoObserver.

In order to keep track of that information in a tidy manner, the new
class RemoteUser has been added.

The UserList now holds a dict where the keys are strings (nicknames)
and the values are RemoteUser objects.